Refactoring, Chrome headless support, Travis CI support (#4)
* Add Travis CI

* Enable Chrome's headless mode

* Try with --no-sandbox

* Refactor the ChromeManager, add a script to download bin

* Throw error when process is terminated.

* debug again

* debug...

* ps ax

* add exec

* Update README.md

+++ src/Client.php
@@ -13,8 +13,6 @@ declare(strict_types=1);
 
 namespace Panthere;
 
-use Facebook\WebDriver\Remote\DesiredCapabilities;
-use Facebook\WebDriver\Remote\RemoteWebDriver;
 use Facebook\WebDriver\WebDriver;
 use Facebook\WebDriver\WebDriverBy;
 use Facebook\WebDriver\WebDriverExpectedCondition;
@@ -22,7 +20,8 @@ use Panthere\Cookie\CookieJar;
 use Panthere\DomCrawler\Crawler;
 use Panthere\DomCrawler\Form as PanthereForm;
 use Panthere\DomCrawler\Link as PanthereLink;
-use Panthere\ProcessManager\ChromeDriver;
+use Panthere\ProcessManager\BrowserManagerInterface;
+use Panthere\ProcessManager\ChromeManager;
 use Symfony\Component\BrowserKit\Client as BaseClient;
 use Symfony\Component\BrowserKit\Request;
 use Symfony\Component\BrowserKit\Response;
@@ -37,53 +36,33 @@ final class Client extends BaseClient implements WebDriver
     use ExceptionThrower;
 
     /**
-     * @var RemoteWebDriver
+     * @var WebDriver
      */
     private $webDriver;
+    private $browserManager;
 
     /**
-     * @var ChromeDriver|null
+     * @param string[]|null $arguments
      */
-    private $chromeDriver;
-
-    private $webDriverFactory;
-
-    public function __construct(?callable $webDriverFactory = null)
+    public static function createChromeClient(?string $chromeDriverBinary = null, ?array $arguments = null): self
     {
-        $this->webDriverFactory = $webDriverFactory;
+        return new self(new ChromeManager($chromeDriverBinary, $arguments));
+    }
+
+    public function __construct(BrowserManagerInterface $browserManager)
+    {
+        $this->browserManager = $browserManager;
+    }
+
+    public function __destruct()
+    {
+        $this->quit();
     }
 
     public function start()
     {
-        if (null === $this->webDriverFactory) {
-            $this->startChromeDriver();
-
-            return;
-        }
-
-        $factory = $this->webDriverFactory;
-        $this->webDriver = $factory();
-    }
-
-    private function startChromeDriver(): void
-    {
-        if (null !== $this->webDriver) {
-            return;
-        }
-
-        if (null === $this->chromeDriver) {
-            $this->chromeDriver = new ChromeDriver();
-            $this->chromeDriver->run();
-        }
-
-        $this->webDriver = RemoteWebDriver::create('http://localhost:9515', DesiredCapabilities::chrome());
-    }
-
-    private function stopChromeDriver(): void
-    {
-        if (null !== $this->chromeDriver) {
-            $this->chromeDriver->stop();
-            $this->chromeDriver = null;
+        if (null === $this->webDriver) {
+            $this->webDriver = $this->browserManager->start();
         }
     }
 
@@ -181,6 +160,7 @@ final class Client extends BaseClient implements WebDriver
 
     protected function createCrawler(): Crawler
     {
+        $this->start();
         $elements = $this->webDriver->findElements(WebDriverBy::cssSelector('html'));
 
         return new Crawler($elements, $this->webDriver, $this->webDriver->getCurrentURL());
@@ -193,6 +173,7 @@ final class Client extends BaseClient implements WebDriver
 
     public function back()
     {
+        $this->start();
         $this->webDriver->navigate()->back();
 
         return $this->crawler = $this->createCrawler();
@@ -200,6 +181,7 @@ final class Client extends BaseClient implements WebDriver
 
     public function forward()
     {
+        $this->start();
         $this->webDriver->navigate()->forward();
 
         return $this->crawler = $this->createCrawler();
@@ -207,6 +189,7 @@ final class Client extends BaseClient implements WebDriver
 
     public function reload()
     {
+        $this->start();
         $this->webDriver->navigate()->refresh();
 
         return $this->crawler = $this->createCrawler();
@@ -219,13 +202,18 @@ final class Client extends BaseClient implements WebDriver
 
     public function restart()
     {
-        $this->webDriver->manage()->deleteAllCookies();
+        if (null !== $this->webDriver) {
+            $this->webDriver->manage()->deleteAllCookies();
+        }
+
         $this->quit();
         $this->start();
     }
 
     public function getCookieJar()
     {
+        $this->start();
+
         return new CookieJar($this->webDriver);
     }
 
@@ -238,6 +226,8 @@ final class Client extends BaseClient implements WebDriver
 
     public function getWebDriver(): WebDriver
     {
+        $this->start();
+
         return $this->webDriver;
     }
 
@@ -248,7 +238,6 @@ final class Client extends BaseClient implements WebDriver
         $this->request = $this->internalRequest = new Request($uri, 'GET');
         $this->webDriver->get($uri);
         $this->response = $this->internalResponse = new Response($this->webDriver->getPageSource());
-
         $this->crawler = $this->createCrawler();
 
         return $this;
@@ -256,83 +245,108 @@ final class Client extends BaseClient implements WebDriver
 
     public function close()
     {
+        $this->start();
+
         return $this->webDriver->close();
     }
 
     public function getCurrentURL()
     {
+        $this->start();
+
         return $this->webDriver->getCurrentURL();
     }
 
     public function getPageSource()
     {
+        $this->start();
+
         return $this->webDriver->getPageSource();
     }
 
     public function getTitle()
     {
+        $this->start();
+
         return $this->webDriver->getTitle();
     }
 
     public function getWindowHandle()
     {
+        $this->start();
+
         return $this->webDriver->getWindowHandle();
     }
 
     public function getWindowHandles()
     {
+        $this->start();
+
         return $this->webDriver->getWindowHandles();
     }
 
     public function quit()
     {
-        if (null === $this->webDriver) {
-            return;
+        if (null !== $this->webDriver) {
+            $this->webDriver->quit();
+            $this->webDriver = null;
         }
-
-        $this->webDriver->quit();
-        $this->webDriver = null;
-
-        $this->stopChromeDriver();
+        $this->browserManager->quit();
     }
 
     public function takeScreenshot($saveAs = null)
     {
+        $this->start();
+
         return $this->webDriver->takeScreenshot($saveAs);
     }
 
     public function wait($timeoutInSecond = 30, $intervalInMillisecond = 250)
     {
+        $this->start();
+
         return $this->webDriver->wait($timeoutInSecond, $intervalInMillisecond);
     }
 
     public function manage()
     {
+        $this->start();
+
         return $this->webDriver->manage();
     }
 
     public function navigate()
     {
+        $this->start();
+
         return $this->webDriver->navigate();
     }
 
     public function switchTo()
     {
+        $this->start();
+
         return $this->webDriver->switchTo();
     }
 
     public function execute($name, $params)
     {
+        $this->start();
+
         return $this->webDriver->execute($name, $params);
     }
 
     public function findElement(WebDriverBy $locator)
     {
+        $this->start();
+
         return $this->webDriver->findElement($locator);
     }
 
     public function findElements(WebDriverBy $locator)
     {
+        $this->start();
+
         return $this->webDriver->findElements($locator);
     }
 }
